+++ packages/components/src/FormField/FormFieldDescription.tsx
@@ -1,10 +1,10 @@
-import React from "react";
+import React, { ReactNode } from "react";
 import styles from "./FormField.module.css";
 import { Text } from "../Text";
 
 interface FormFieldDescriptionProps {
   readonly id: string;
-  readonly description?: string;
+  readonly description?: ReactNode;
   readonly visible?: boolean;
 }
 
@@ -15,11 +15,17 @@ export function FormFieldDescription({
 }: FormFieldDescriptionProps) {
   if (!visible) return null;
 
+  const useStringFormat = !description || typeof description === "string";
+
   return (
     <div id={id} className={styles.description}>
-      <Text size="small" variation="subdued">
-        {description}
-      </Text>
+      {useStringFormat ? (
+        <Text size="small" variation="subdued">
+          {description}
+        </Text>
+      ) : (
+        description
+      )}
     </div>
   );
 }